l1-pkb/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sentence-transformers==3.3.1
torch>=2.0.0
qdrant-client==1.12.1
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
watchdog>=4.0.0
pyyaml>=6.0
